Algunos cambios hechos

+++ SG_MAUI_RamSerDav_/MVVM/ViewModels/GestionUsuariosViewModel.cs
@@ -3,7 +3,9 @@ using SG_MAUI_RamSerDav_.MVVM.Models;
 using SG_MAUI_RamSerDav_.MVVM.Views;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
@@ -19,10 +21,12 @@ namespace SG_MAUI_RamSerDav_.MVVM.ViewModels
         public ICommand btnSalirCommand { get; set; }
         public ICommand EliminarCommand { get; set; }
         public ICommand GuardarCommand { get; set; }
+        
 
         public GestionUsuariosViewModel()
         {
             refrescarLista();
+
             btnIrGestionUsuariosCommand = new Command(() =>
             {
                 App.Current.MainPage.Navigation.PushAsync(new GestionUsuariosView());
@@ -32,27 +36,52 @@ namespace SG_MAUI_RamSerDav_.MVVM.ViewModels
 
             btnSalirCommand = new Command(() =>
             {
-                //Vuelve a la página del login pero habría que pensar si se cierra la aplicación o no
+                //Vuelve a la página del login
                 App.Current.MainPage.Navigation.PopAsync();
+                // Si quieres salir de la aplicación completamente
+                // App.Current.Quit();
             });
 
-            EliminarCommand = new Command(() =>
+
+            EliminarCommand = new Command(async () =>
             {
-                App.UsuarioRepo.DeleteItem(UsuarioActual);
-                refrescarLista();
+                bool resultado = await Auxiliar.Herramientas.MensajeConfirmacion("Confirmación", "¿Estás seguro?");
+                if (resultado)
+                { 
+                    App.UsuarioRepo.DeleteItem(UsuarioActual);
+                    refrescarLista();
+                }       
             });
+
 
             GuardarCommand = new Command(() =>
             {
-                App.UsuarioRepo.SaveItem(UsuarioActual);
-                Console.WriteLine(App.UsuarioRepo.StatusMessage);
-                refrescarLista();
-                nuevoUsuario();
+                if (camposCompletos())
+                {
+                    App.UsuarioRepo.SaveItemCascade(UsuarioActual);
+                    refrescarLista();
+                    limpiarCampos();
+                }
+                else
+                {
+                    Auxiliar.Herramientas.MensajeInfomativo("Faltan campos por completar \n o usuario por seleccionar");
+                }
             });
 
         }
 
-        public void nuevoUsuario()
+        //public void checkbox_chaged()
+        //{
+        //    UsuarioActual.EsDelegado = !UsuarioActual.EsDelegado;
+        //}
+
+        private bool camposCompletos()
+        {
+            return !string.IsNullOrWhiteSpace(UsuarioActual.Email) && !string.IsNullOrWhiteSpace(UsuarioActual.Password);
+        }
+
+
+        public void limpiarCampos()
         {
             UsuarioActual = new Usuario();
         }
